Improve and reduce some code in Installer
