+++ core/common/src/main/java/com/looker/core/common/extension/Collections.kt
@@ -1,0 +1,5 @@
+package com.looker.core.common.extension
+
+fun <T> Set<T>.updateAsMutable(block: MutableSet<T>.() -> Unit): Set<T> {
+	return toMutableSet().apply(block)
+}
