+++ installer/src/main/java/com/looker/installer/model/InstallerQueueState.kt
@@ -1,0 +1,16 @@
+package com.looker.installer.model
+
+data class InstallerQueueState(
+	val currentItem: InstallItemState,
+	val queued: Set<String>
+) {
+	companion object {
+		val EMPTY = InstallerQueueState(InstallItemState.EMPTY, emptySet())
+	}
+}
+
+infix fun String.isQueuedIn(state: InstallerQueueState): Boolean = this in state.queued
+
+infix fun String.isInstalling(state: InstallerQueueState): Boolean =
+	this == state.currentItem.installedItem.packageName.name
+			&& state.currentItem.state is InstallState.Installing
